// food-channel-processor/src/v1/channel-order-food/model/channel-food-sumary-data.data.model.ts
export class ChannelFoodSumaryDataModel {

    commission_amount_SHF: number;
    commission_amount_GRF: number;
    commission_amount_BEF: number;
    total_commission_amount: number;
    order_amount_SHF: number;
    order_amount_GRF: number;
    order_amount_BEF: number;
    total_order_amount: number;
    total_amount_SHF: number;
    total_amount_GRF: number;
    total_amount_BEF: number;
    total_revenue_amount: number;
    discount_amount_SHF: number;
    discount_amount_GRF: number;
    discount_amount_BEF: number;
    total_discount_amount: number;

}